--- src/categories/main_blocks.ts
@@ -1,21 +1,0 @@
-import "../blocks/structure";
-
-export const MainCategory = {
-    kind: "category",
-    name: "Web Structure",
-    colour: "#5CA699",
-    contents: [
-      {
-        kind: "block",
-        type: "html_tag",
-      },
-      {
-        kind: "block",
-        type: "head_tag",
-      },
-      {
-        kind: "block",
-        type: "body_tag",
-      },
-    ],
-  };
