+++ src/workspaces/frontend/frontendWorkspace.tsx
@@ -5,7 +5,6 @@ import { javascriptCategory } from "../../categories/javascript";
 import { htmlCategory } from "../../categories/html";
 import { commonCategory } from "../../categories/google_blocks"
 import { cssCategory } from "../../categories/css"; 
-import { MainCategory } from "../../categories/main_blocks"; 
 
 
 interface FrontendWorkspaceProps {
@@ -18,7 +17,7 @@ function FrontendWorkspace({ onCodeChange }: FrontendWorkspaceProps) {
   const toolboxCategories = {
     kind: "categoryToolbox",
 
-    contents: [MainCategory, htmlCategory, cssCategory, javascriptCategory, commonCategory],
+    contents: [htmlCategory, cssCategory, javascriptCategory, commonCategory],
   };
 
   const workspaceDidChange = (workspace: WorkspaceSvg) => {
